leam-base.joda-time.mid-2424.idx-177786.6.mutant

+++ joda-time/src/main/java/org/joda/time/Period.java
@@ -776,7 +776,7 @@ public final class Period
      * @return the number of weeks in the period, zero if unsupported
      */
     public int getWeeks() {
-        return getPeriodType().getIndexedField(this, PeriodType.WEEK_INDEX);
+        return (int) getPeriodType().getIndexedField(this, PeriodType.WEEK_INDEX);
     }
 
     /**
